Working on A2 some more

+++ src/edu/drexel/cs430/renderengine/shapes/Point.java
@@ -4,27 +4,27 @@ package edu.drexel.cs430.renderengine.shapes;
  * Created by Angel on 9/25/2016.
  */
 public class Point {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
